Update letter_sum_methods.py
Dramatically changed the method to find the mode score in a list of words. The code is much easier to read and concise after importing mode from the statistics class, and using a list comprehension to add each half of the master life into a complete list after computing values for each list on their own.